+++ src/DOMStringMap.php
@@ -12,6 +12,24 @@ namespace Wikimedia\IDLeDOM;
  *
  * @phan-forbid-undeclared-magic-properties
  */
-interface DOMStringMap {
+interface DOMStringMap extends \ArrayAccess {
+	/**
+	 * @param string $name
+	 * @return string
+	 */
+	public function namedItem( string $name ) : string;
+
+	/**
+	 * @param string $name
+	 * @param string $value
+	 * @return void
+	 */
+	public function setNamedItem( string $name, string $value ) : void;
+
+	/**
+	 * @param string $name
+	 * @return void
+	 */
+	public function removeNamedItem( string $name ) : void;
 
 }